chore: use plugin way to expose seever info endpoint
