+++ apisix/admin/init.lua
@@ -30,7 +30,6 @@ local reload_event = "/apisix/admin/plugins/reload"
 local ipairs = ipairs
 local error = error
 local events
-local server_info
 
 local MAX_REQ_BODY = 1024 * 1024 * 1.5      -- 1.5 MiB
 
@@ -261,17 +260,6 @@ local function post_reload_plugins()
 end
 
 
-local function get_server_info()
-    local server_info, err = server_info.get()
-    if not server_info then
-        core.log.error("failed to get server_info: ", err)
-        core.response.exit(500, err)
-    end
-
-    core.response.exit(200, server_info)
-end
-
-
 local function sync_local_conf_to_etcd()
     core.log.warn("sync local conf to etcd")
 
@@ -330,11 +318,6 @@ local uri_route = {
         methods = {"PUT"},
         handler = post_reload_plugins,
     },
-    {
-        paths = [[/apisix/admin/server_info]],
-        methods = {"GET"},
-        handler = get_server_info,
-    }
 }
 
 
@@ -343,8 +326,6 @@ function _M.init_worker()
     if not local_conf.apisix or not local_conf.apisix.enable_admin then
         return
     end
-
-    server_info = require("apisix.server_info")
 
     router = route.new(uri_route)
     events = require("resty.worker.events")
